requirements.txt
# ============================================================================
# FIAP Fase 7 - Sistema de Gestão para Agronegócio
# Requirements.txt consolidado de todas as fases
# ============================================================================

# Core
python-dotenv==1.0.0
pyyaml==6.0.1

# Web Framework & Dashboard
streamlit==1.29.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
flask==3.0.0
flask-cors==4.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Data Analysis & Processing
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
joblib==1.3.2

# Computer Vision & Deep Learning
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.206
opencv-python==4.8.1.78
pillow==10.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# AWS Services
boto3==1.29.7
awscli==1.30.7

# HTTP & API
requests==2.31.0
urllib3==2.1.0
aiohttp==3.9.1

# IoT/Sensors & MQTT
paho-mqtt==1.6.1

# Data Validation
pydantic==2.5.2

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Environment & Configuration
configparser==6.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
pylint==3.0.3

# Documentation
sphinx==7.2.6

# Utilities
tqdm==4.66.1
python-json-logger==2.0.7

# Email & Templates
jinja2==3.1.2
